+++ release.sh
@@ -6,15 +6,14 @@ rm -Rf coverage
 rm -Rf dist
 rm -Rf node_modules
 rm junit.xml
-rm yarn.lock
+rm pnpm-lock.yaml
 
 # Prepare
-yarn install
-yarn lint
-yarn test --silent
-yarn audit --groups dependencies
-yarn audit --groups peerDependencies
-yarn build
+pnpm install
+pnpm run lint
+pnpm run test -- --silent
+pnpm audit
+pnpm run build
 
 # Publish
 npm publish --dry-run