+++ src/options/pug-class-notation.ts
@@ -10,9 +10,10 @@ export const PUG_CLASS_NOTATION: ChoiceSupportOption = {
 	description: 'Define how classes should be formatted.',
 	choices: [
 		{ value: 'literal', description: 'Forces all valid classes to be printed as literals.' },
+		{ value: 'attribute', description: 'Forces all valid classes to be printed in class attribute.' },
 		{ value: 'as-is', description: 'Disables class formatting.' }
 	]
 };
 
 /** Pug class notation. */
-export type PugClassNotation = 'literal' | 'as-is';
+export type PugClassNotation = 'literal' | 'attribute' | 'as-is';