+++ src/options/pug-framework.ts
@@ -2,14 +2,14 @@ import type { ChoiceSupportOption } from 'prettier';
 import { CATEGORY_PUG } from '.';
 
 /** Pug Framework. */
-export const PUG_FRAMEWORK: ChoiceSupportOption = {
+export const PUG_FRAMEWORK: ChoiceSupportOption<PugFramework> = {
 	since: '1.14.0',
 	category: CATEGORY_PUG,
 	type: 'choice',
-	default: 'none',
+	default: 'auto',
 	description: 'Define which framework is used in the project.',
 	choices: [
-		{ value: 'none', description: 'No framework specifically.' },
+		{ value: 'auto', description: 'Try to identify used framework, if there is one.' },
 		{ value: 'vue', description: 'Uses Vue.js.' },
 		{ value: 'svelte', description: 'Uses Svelte.' },
 		{ value: 'angular', description: 'Uses Angular.' }
@@ -17,4 +17,4 @@ export const PUG_FRAMEWORK: ChoiceSupportOption = {
 };
 
 /** Pug class notation. */
-export type PugFramework = 'none' | 'vue' | 'svelte' | 'angular';
+export type PugFramework = 'auto' | 'vue' | 'svelte' | 'angular';