+++ src/utils/angular.ts
@@ -1,3 +1,5 @@
+import { isQuoted, isWrappedWith } from './common';
+
 /**
  * Indicates whether the attribute name is an Angular binding.
  *
@@ -79,13 +81,5 @@ export function isAngularDirective(name: string): boolean {
  * @returns `true` if `val` passes the angular interpolation check, otherwise `false`.
  */
 export function isAngularInterpolation(val: string): boolean {
-	return (
-		val.length >= 5 &&
-		((val[0] === '"' && val[val.length - 1] === '"') || (val[0] === "'" && val[val.length - 1] === "'")) &&
-		val[1] === '{' &&
-		val[2] === '{' &&
-		val[val.length - 2] === '}' &&
-		val[val.length - 3] === '}' &&
-		!val.includes('{{', 3)
-	);
+	return val.length >= 5 && isQuoted(val) && isWrappedWith(val, '{{', '}}', 1) && !val.includes('{{', 3);
 }