+++ tests/issues/issue-262/issue-262.test.ts
@@ -1,0 +1,18 @@
+import { readFileSync } from 'fs';
+import { resolve } from 'path';
+import { format } from 'prettier';
+import { plugin } from './../../../src/index';
+
+describe('Issues', () => {
+	test('should format inline css styles', () => {
+		const expected: string = readFileSync(resolve(__dirname, 'formatted.pug'), 'utf8');
+		const code: string = readFileSync(resolve(__dirname, 'unformatted.pug'), 'utf8');
+		const actual: string = format(code, {
+			parser: 'pug',
+			plugins: [plugin],
+			useTabs: true
+		});
+
+		expect(actual).toBe(expected);
+	});
+});