+++ tests/issues/issue-282/id-indented.vue
@@ -1,0 +1,14 @@
+<template lang="pug">
+  #id
+    div
+      p Lorem ipsum dolor sit amet, consectetur adipisici elit
+    div
+      div
+        p Lorem ipsum dolor sit amet, consectetur adipisici elit
+</template>
+
+<script>
+export default {
+  name: "Hello",
+};
+</script>